MISC: Rename example 2 runscript
